refactor: move runtime config types to kit
