+++ docs/content/3.docs/1.usage/5.runtime-config.md
@@ -48,10 +48,8 @@ export default async () => {
 
 It is possible to manually type your runtime config.
 
-```ts [nuxt.config.js|ts]
-import { defineNuxtConfig } from 'nuxt3'
-
-declare module '@nuxt/nitro' {
+```ts [index.d.ts]
+declare module '@nuxt/kit' {
   interface PublicRuntimeConfig {
     testConfig: string
   }
@@ -60,12 +58,7 @@ declare module '@nuxt/nitro' {
   }
 }
 
-export default defineNuxtConfig({
-  privateRuntimeConfig: {
-    token: process.env.SECRET_TOKEN
-  },
-  publicRuntimeConfig: {
-    testConfig: 'test'
-  }
-})
+// It is always important to ensure you import/export
+// something when augmenting a type
+export {}
 ```
