+++ packages/kit/src/index.ts
@@ -25,3 +25,4 @@ export * from './types/module'
 export * from './types/nuxt'
 export * from './types/components'
 export * from './types/imports'
+export * from './types/runtime-config'
