+++ packages/kit/src/types/runtime-config.ts
@@ -1,0 +1,2 @@
+export interface PublicRuntimeConfig extends Record<string, any> { }
+export interface PrivateRuntimeConfig extends PublicRuntimeConfig { }
